rename make-primative to correct typo
